Monitoring: Implement endpoint for CSV export
New endpoints:

- /monitoring/v1/export/download/
Responds with a string that contains all the csv lines.
Start time and end time are required fields.
Create a custom CSVRenderer for this endpoints.
When it comes to filtering data with the time we do it by having
a time range where any parking that was active during
the range gets exported. This means that any parking
that started before range and ended before range OR started
after range will not get filtered in.
One important thing to mention is that if the queryset
ends up being very big (>100k objs) ram usage may suffer
and the user will have to wait a long time for a response from
the server since we don't stream the response.