Monitor: CSVRenderer exception handling
In a case of an action that uses the CSVRenderer
raises an exception, change the renderer to a
JSONRenderer so the server can respond with an
error message with correct data format.

We have to bypass the custom CSVRenderer because
otherwise the validation error message gets passed
over to the CSVRenderer and the CSVRenderer can't
really do anything with the error message.